Support the Sitecore 7.5 zip archive format.
Properly extract Sitecore source files from zip files which have a
different file structure depending on which version of the CMS is being
used.

Zip archives from SDN of Sitecore 7.2+ nest the website files in a
containing folder, named for the major and revision version numbers of
the CMS. This is different from the 7.0 distribution, which has the
Data, Databases, and Website folders at the root of the archive.